po -> po_thumbnailer
po is a "reserved" directory name for scripty since in 99.99% of cases
if you have a po folder means you are storing the po in your repo
which is not the correct thing to do. Here is it correct but for the 0.01% of the
cases i'm just renaming the folder.